convex comb

+++ conv_comb/conv_comb.cpp
@@ -280,8 +280,8 @@ signed main(int argc, char const *argv[]){
 			double wrst = 0;
 			int id = 0;
 			for (int i = 1; i < fx.getDim(); i++){
-				int x = fx[i];
-				fx.point[i] = max(0.0	, fx[i] - 0.5);
+				double x = fx[i];
+				fx.point[i] = max(0.0, fx[i] - 0.5);
 
 				cout<<'\t'<<i<<' '<<fx<<endl;
 				double aux = ConvexComb(sol, fx, int_points);
